# artifacts/api-server/src/routes/directReports.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, directReportsTable, organizationsTable, activityTable } from "@workspace/db";
import {
  CreateDirectReportBody,
  GetDirectReportParams,
  GetDirectReportResponse,
  UpdateDirectReportParams,
  UpdateDirectReportBody,
  UpdateDirectReportResponse,
  DeleteDirectReportParams,
  ListDirectReportsResponse,
} from "@workspace/api-zod";

const router: IRouter = Router();

const directReportSelect = {
  id: directReportsTable.id,
  name: directReportsTable.name,
  role: directReportsTable.role,
  email: directReportsTable.email,
  phone: directReportsTable.phone,
  organizationId: directReportsTable.organizationId,
  organizationName: organizationsTable.name,
  status: directReportsTable.status,
  hireDate: directReportsTable.hireDate,
  performanceRating: directReportsTable.performanceRating,
  avatarUrl: directReportsTable.avatarUrl,
  createdAt: directReportsTable.createdAt,
};

function mapNulls(r: any) {
  return {
    ...r,
    phone: r.phone ?? undefined,
    organizationId: r.organizationId ?? undefined,
    organizationName: r.organizationName ?? undefined,
    hireDate: r.hireDate ?? undefined,
    performanceRating: r.performanceRating ?? undefined,
    avatarUrl: r.avatarUrl ?? undefined,
  };
}

router.get("/direct-reports", async (_req, res): Promise<void> => {
  const reports = await db
    .select(directReportSelect)
    .from(directReportsTable)
    .leftJoin(organizationsTable, eq(directReportsTable.organizationId, organizationsTable.id))
    .orderBy(directReportsTable.name);
  res.json(ListDirectReportsResponse.parse(reports.map(mapNulls)));
});

router.post("/direct-reports", async (req, res): Promise<void> => {
  const parsed = CreateDirectReportBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [report] = await db
    .insert(directReportsTable)
    .values(parsed.data)
    .returning();

  await db.insert(activityTable).values({
    type: "report_added",
    message: `${report.name} was added as a direct report`,
    entityName: report.name,
  });

  const [full] = await db
    .select(directReportSelect)
    .from(directReportsTable)
    .leftJoin(organizationsTable, eq(directReportsTable.organizationId, organizationsTable.id))
    .where(eq(directReportsTable.id, report.id));

  res.status(201).json(GetDirectReportResponse.parse(mapNulls(full)));
});

router.get("/direct-reports/:id", async (req, res): Promise<void> => {
  const params = GetDirectReportParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [report] = await db
    .select(directReportSelect)
    .from(directReportsTable)
    .leftJoin(organizationsTable, eq(directReportsTable.organizationId, organizationsTable.id))
    .where(eq(directReportsTable.id, params.data.id));

  if (!report) {
    res.status(404).json({ error: "Direct report not found" });
    return;
  }

  res.json(GetDirectReportResponse.parse(mapNulls(report)));
});

router.patch("/direct-reports/:id", async (req, res): Promise<void> => {
  const params = UpdateDirectReportParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateDirectReportBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [updated] = await db
    .update(directReportsTable)
    .set(parsed.data)
    .where(eq(directReportsTable.id, params.data.id))
    .returning();

  if (!updated) {
    res.status(404).json({ error: "Direct report not found" });
    return;
  }

  const [full] = await db
    .select(directReportSelect)
    .from(directReportsTable)
    .leftJoin(organizationsTable, eq(directReportsTable.organizationId, organizationsTable.id))
    .where(eq(directReportsTable.id, params.data.id));

  res.json(UpdateDirectReportResponse.parse(mapNulls(full)));
});

router.delete("/direct-reports/:id", async (req, res): Promise<void> => {
  const params = DeleteDirectReportParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [report] = await db
    .delete(directReportsTable)
    .where(eq(directReportsTable.id, params.data.id))
    .returning();

  if (!report) {
    res.status(404).json({ error: "Direct report not found" });
    return;
  }

  res.sendStatus(204);
});

export default router;