# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental CEO Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DailyTop3Item {
  id: number;
  title: string;
  description?: string;
  completed: boolean;
  priority: number;
  date: string;
  createdAt: string;
}

export interface CreateDailyTop3Body {
  title: string;
  description?: string;
  priority: number;
}

export interface UpdateDailyTop3Body {
  title?: string;
  description?: string;
  completed?: boolean;
  priority?: number;
}

export type OrganizationStatus =
  (typeof OrganizationStatus)[keyof typeof OrganizationStatus];

export const OrganizationStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Organization {
  id: number;
  name: string;
  address: string;
  city: string;
  state: string;
  phone?: string;
  email?: string;
  providerCount?: number;
  patientCount?: number;
  monthlyRevenue?: number;
  status: OrganizationStatus;
  createdAt: string;
}

export type CreateOrganizationBodyStatus =
  (typeof CreateOrganizationBodyStatus)[keyof typeof CreateOrganizationBodyStatus];

export const CreateOrganizationBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateOrganizationBody {
  name: string;
  address: string;
  city: string;
  state: string;
  phone?: string;
  email?: string;
  providerCount?: number;
  patientCount?: number;
  monthlyRevenue?: number;
  status?: CreateOrganizationBodyStatus;
}

export type UpdateOrganizationBodyStatus =
  (typeof UpdateOrganizationBodyStatus)[keyof typeof UpdateOrganizationBodyStatus];

export const UpdateOrganizationBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateOrganizationBody {
  name?: string;
  address?: string;
  city?: string;
  state?: string;
  phone?: string;
  email?: string;
  providerCount?: number;
  patientCount?: number;
  monthlyRevenue?: number;
  status?: UpdateOrganizationBodyStatus;
}

export type DirectReportStatus =
  (typeof DirectReportStatus)[keyof typeof DirectReportStatus];

export const DirectReportStatus = {
  active: "active",
  on_leave: "on_leave",
  inactive: "inactive",
} as const;

export interface DirectReport {
  id: number;
  name: string;
  role: string;
  email: string;
  phone?: string;
  organizationId?: number;
  organizationName?: string;
  status: DirectReportStatus;
  hireDate?: string;
  performanceRating?: number;
  avatarUrl?: string;
  createdAt: string;
}

export type CreateDirectReportBodyStatus =
  (typeof CreateDirectReportBodyStatus)[keyof typeof CreateDirectReportBodyStatus];

export const CreateDirectReportBodyStatus = {
  active: "active",
  on_leave: "on_leave",
  inactive: "inactive",
} as const;

export interface CreateDirectReportBody {
  name: string;
  role: string;
  email: string;
  phone?: string;
  organizationId?: number;
  status?: CreateDirectReportBodyStatus;
  hireDate?: string;
  performanceRating?: number;
  avatarUrl?: string;
}

export type UpdateDirectReportBodyStatus =
  (typeof UpdateDirectReportBodyStatus)[keyof typeof UpdateDirectReportBodyStatus];

export const UpdateDirectReportBodyStatus = {
  active: "active",
  on_leave: "on_leave",
  inactive: "inactive",
} as const;

export interface UpdateDirectReportBody {
  name?: string;
  role?: string;
  email?: string;
  phone?: string;
  organizationId?: number;
  status?: UpdateDirectReportBodyStatus;
  hireDate?: string;
  performanceRating?: number;
  avatarUrl?: string;
}

export type AnnouncementType =
  (typeof AnnouncementType)[keyof typeof AnnouncementType];

export const AnnouncementType = {
  info: "info",
  warning: "warning",
  success: "success",
  urgent: "urgent",
} as const;

export interface Announcement {
  id: number;
  title: string;
  content: string;
  type: AnnouncementType;
  createdAt: string;
}

export type CreateAnnouncementBodyType =
  (typeof CreateAnnouncementBodyType)[keyof typeof CreateAnnouncementBodyType];

export const CreateAnnouncementBodyType = {
  info: "info",
  warning: "warning",
  success: "success",
  urgent: "urgent",
} as const;

export interface CreateAnnouncementBody {
  title: string;
  content: string;
  type: CreateAnnouncementBodyType;
}

export interface DashboardSummary {
  totalOrganizations: number;
  totalDirectReports: number;
  totalPatients: number;
  totalMonthlyRevenue: number;
  dailyTop3Completed: number;
  dailyTop3Total: number;
  activeOrganizations: number;
  revenueChangePercent?: number;
  patientChangePercent?: number;
}

export interface OrgPerformance {
  organizationId: number;
  organizationName: string;
  monthlyRevenue: number;
  patientCount: number;
  providerCount: number;
  productionGoalPercent: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  task_completed: "task_completed",
  report_added: "report_added",
  org_updated: "org_updated",
  announcement: "announcement",
  milestone: "milestone",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  entityName?: string;
  createdAt: string;
}

export type GetRecentActivityParams = {
  limit?: number;
};