# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental CEO Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementType } from "./announcementType";

export interface Announcement {
  id: number;
  title: string;
  content: string;
  type: AnnouncementType;
  createdAt: Date;
}