# lib/api-zod/src/generated/types/createDirectReportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental CEO Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDirectReportBodyStatus } from "./createDirectReportBodyStatus";

export interface CreateDirectReportBody {
  name: string;
  role: string;
  email: string;
  phone?: string;
  organizationId?: number;
  status?: CreateDirectReportBodyStatus;
  hireDate?: Date;
  performanceRating?: number;
  avatarUrl?: string;
}