# lib/api-zod/src/generated/types/createOrganizationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental CEO Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrganizationBodyStatus } from "./createOrganizationBodyStatus";

export interface CreateOrganizationBody {
  name: string;
  address: string;
  city: string;
  state: string;
  phone?: string;
  email?: string;
  providerCount?: number;
  patientCount?: number;
  monthlyRevenue?: number;
  status?: CreateOrganizationBodyStatus;
}